refactor: extract DynascaleManager
