+++ packages/client/index.ts
@@ -18,9 +18,10 @@ export * from './src/StreamSfuClient';
 export * from './src/devices';
 export * from './src/store';
 export * from './src/sorting';
+export * from './src/helpers/DynascaleManager';
 export * from './src/helpers/ViewportTracker';
-
 export * from './src/helpers/sound-detector';
 export * as Browsers from './src/helpers/browsers';
+
 export * from './src/client-details';
 export * from './src/logger';
