+++ packages/react-sdk/src/core/components/CallLayout/SpeakerLayout.tsx
@@ -57,8 +57,7 @@ export const SpeakerLayout = ({
   useEffect(() => {
     if (!scrollWrapper || !call) return;
 
-    const cleanup = call.viewportTracker.setViewport(scrollWrapper);
-
+    const cleanup = call.setViewport(scrollWrapper);
     return () => cleanup();
   }, [scrollWrapper, call]);
